edit

+++ ERP_System.Tables/Guide/Invoice.cs
@@ -37,9 +37,9 @@ namespace ERP_System.Tables
         public string? StockName { get; set; }
         public Guid? StockId { get; set; }
 
-        public double? Qty { get; set; }
-        public double? PricePerUnit { get; set; }
-        public double? PricePerQty  => Qty * PricePerUnit;
+        public decimal? Qty { get; set; }
+        public decimal? PricePerUnit { get; set; }
+        public decimal? PricePerQty  => Qty * PricePerUnit;
 
 
         [ForeignKey(nameof(Invoice))]
